+++ h2integrate/core/model_base.py
@@ -2,6 +2,18 @@ import openmdao.api as om
 
 
 class CostModelBaseClass(om.ExplicitComponent):
+    """Baseclass to be used for all cost models. The built-in outputs
+    are used by the finance model and must be
+
+    Outputs:
+        - CapEx (float): capital expenditure costs in $
+        - OpEx (float): annual operating expenditure costs in $/year
+
+    Discrete Outputs:
+        - cost_year (int): dollar-year corresponding to CapEx and OpEx values.
+            This may be inherent to the cost model, or may depend on user provided input values.
+    """
+
     def initialize(self):
         self.options.declare("driver_config", types=dict)
         self.options.declare("plant_config", types=dict)
@@ -11,6 +23,7 @@ class CostModelBaseClass(om.ExplicitComponent):
         # Define outputs: CapEx and OpEx costs
         self.add_output("CapEx", val=0.0, units="USD", desc="Capital expenditure")
         self.add_output("OpEx", val=0.0, units="USD/year", desc="Operational expenditure")
+        # Define discrete outputs: cost_year
         self.add_discrete_output("cost_year", val=0, desc="Dollar year for costs")
 
     def compute(self, inputs, outputs, discrete_inputs, discrete_outputs):